The basic start
This has a function(2 technically) to pull historical data from Kraken API, using the Requests (http://docs.python-requests.org/en/master/user/quickstart/) module for python. The Kraken data is not in most convenient form, so there's a function to convert it into a more easy to handle and plot format. There is a function to calculate simple moving average of an array. Those are the main ones. There's also a bit at the bottom that tests a simple turning point of moving average strategy, but that is messy.